+++ public/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <script src="http://localhost:8098"></script>
+    <!-- <script src="http://localhost:8098"></script> -->
     <meta charset="utf-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width,initial-scale=1.0" />
@@ -11,8 +11,8 @@
   <body>
     <noscript>
       <strong
-        >We're sorry but project_chrome doesn't work properly without JavaScript enabled. Please
-        enable it to continue.</strong
+        >We're sorry but project_chrome doesn't work properly without JavaScript
+        enabled. Please enable it to continue.</strong
       >
     </noscript>
     <div id="app"></div>
