+++ src/manifest.json
@@ -10,7 +10,7 @@
     "128": "icons/128.png"
   },
   "permissions": ["tabs", "identity"],
-  "content_security_policy": "script-src 'self' https://www.gstatic.com/ https://*.firebaseio.com https://www.googleapis.com; object-src 'self'",
+  "content_security_policy": "script-src 'self' 'unsafe-eval' https://www.gstatic.com/ https://*.firebaseio.com https://www.googleapis.com https://apis.google.com; object-src 'self'",
   "oauth2": {
     "client_id": "876962566748-p2ph7khp7kcml7j2j6gfhe1htsv8m6jj.apps.googleusercontent.com",
     "scopes": [
