+++ src/standalone/standalone.js
@@ -1,4 +1,5 @@
 import Vue from 'vue';
+import firebase from 'firebase';
 import App from './App';
 import store from '../store';
 import router from './router';
@@ -10,3 +11,10 @@ new Vue({
   router,
   render: h => h(App),
 });
+
+const config = {
+  apiKey: 'AIzaSyD8RTIS1VMI76VwdvHhba9wXFyIkbg4PiE',
+  databaseURL: 'https://project-webextension-tg-4dfca.firebaseio.com',
+  projectId: 'project-webextension-tg-4dfca',
+};
+firebase.initializeApp(config);
